Only expose a simple FormatVerilog() function, hide implementation class.
FormatVerilog() absorbed much of the code that was in main()
in verilog_format.cc

Removed/ported tests that mucked around with class internals.
These used to serve some value when we wanted to test the formatter
in isolation from the analyzer by feeding it in processed data,
but maintaining these are more burdensome than their worth, now that the
formatter has matured more.

Note: Formatter test cases are all now run through the (internal)
VerifyFormatting() function.
PiperOrigin-RevId: 292398789