Add build version based on git-hash. If that is not available, use just the timestamp that is always available in bazel.
Given that bazel does not allow to arbitrarily generate changing
code to guarantee hermetic builds, we have to use the workspace status
feature
 https://docs.bazel.build/versions/master/user-manual.html#workspace_status

In short, we create name-value pairs using the build-version.sh script, which
are then picked up in the version_header genrule.

To make sure the build-version.sh script is invoked, we add it to the
project .bazelrc.

This now creates outputs like these

$ bazel-bin/verilog/tools/syntax/verilog_syntax --version
verible_2020-02-13_27b5698

Fixes #173

Signed-off-by: Henner Zeller <h.zeller@acm.org>