Disable module port formatting until alignment is implemented.
It was decided that users would rather leave their beautifully aligned ports
alone until the formatter implements alignment.

The default of verilog::FormatStyle is still format-everything, however,
the verilog_format binary uses a flag with a default of disabling module ports.

Pre-existing tests still run with module ports formatted, but we've added
a small section of tests to cover the new disabling behavior.

PiperOrigin-RevId: 301465346