PR #232: Fix dynamic_cast failure on [create-object-name-match] rule Closes #222 Closes #172
The comment in the file actually says:

```
// Returns token information for a single string literal expression, or nullptr
// if the expression is not a string literal.
// `expr_node` should be a SyntaxTreeNode tagged as an expression.
```

So this extra check in fact verifies if the given token can even be a string literal.

The problem in #222 was that the argument was in fact an expression (called as `l("name")` instead of just `"name"`) and the present checks failed to fully verify that.

GitHub PR https://github.com/google/verible/pull/232

Copybara import of the project:

  - 456c475544e8ff7228b10b1d83aad5da7fe22655 Fix create object name match rule checks by Tomasz Gorochowik <tgorochowik@antmicro.com>
  - bc60c24b410a051b1874023e847528da62536102 Add more create object name match rule unit tests by Tomasz Gorochowik <tgorochowik@antmicro.com>

Closes #232

PiperOrigin-RevId: 301876327