Signal lexical errors during recursive lex-diff.
Diff function now returns a DiffStatus enum instead of a bool to communicate
the nature of the error condition.

VerilogLexicallyEquivalent() is now formulated in terms of the
now-entirely-language-agnostic (functional programming style)
LexicallyEquivalent().
Recursion is no longer hard-coded, but controlled with a std::function.
There is now no need for separate compare-tokens versions of the equivalence
checking function.

Added more test checks to improve raw coverage and mutation coverage.

PiperOrigin-RevId: 301878111