Enable heterogeneous lookup in C++14 or higher, else fallback to temporary key.
ForwardReferenceElseConstruct is a helper class to forward a straight reference
when the source/destination types match, else construct a temporary rvalue.
This is a fallback mechanism that allows an interface that forwards to
take advantage of heterogeneous lookup (HL) still work with a temporary object
when HL is unavailable, as is the case with pre-C++14 STL containers.

PiperOrigin-RevId: 304209707