Add a simple way to ignore #-comments in lint configuration files.
Lint rules can be configured by flag, in which they are a comma-separated
list of rules. In a configuration file, we have multiple lines with
one rule configured per line. In such a file, it is desirable to allow
for comments. This allows for common end-of-line hash comments.

While at it, allow whitespace before and after rules.

Fixes #274

PiperOrigin-RevId: 309871733