Changing rule strategy and adding test improvements.
In this implementation the rule will handle leafs when context is inside one of the following:
 * kStatement
 * kMacroCall - for when MacroCalls are in class, file, package scope.
 * kDataDeclaration - for when plain Macros are in class, file, package scope.

A two state FSM is implemented to track tokens. In kNormal state it will try to match any
MacroCallId, MacroIdItem or MacroIdentifier token, then chek if it matches '`uvm_'. If it
does, save the mathing token for report and move to kCheckMacro state. Here it will wait for
MacroCallCloseToEndLine (no violation) or ';' which will register a violation, then reset
the state to kNormal. The state is also cleared when outside aforementioned context.