Rename tools prefix with verible-.
The tool names used to be overly generic ('verilog_lint'), so
add the verible prefix to disambiguate them from other tools.
Also use dashes instead of underscores because they look prettier :)

Since tools are intended to possibly support more languages in
the future, we continue with also including 'verilog' in it.
(Decision was to keep the language name in the tool rather than
have a flag choose the language.)

Names are now verible-verilog-lint, verible-verilog-format, ...

For backward compatibility: the old names are still present
for now: there are short shell-scripts with the old name
that invoke the new tools for a transition period
(initially silently, later they might start to write
a warning message to stderr before being decommissioned
after some time).

Fixes: https://github.com/google/verible/issues/169
PiperOrigin-RevId: 315524811