StringSpliterator is a functor that returns tokens one at a time.
It splits, and works like a generator function.

For comparison, absl::StrSplit() returns all results at once, and stores them in
a temporary splitter object that is convertible to a container.
This interface can use constant memory, and can be adapted to fixed size
structures like std::array, std::tuple.

PiperOrigin-RevId: 317221920