Patch (unified-diff) reader library, using top-level PatchSet structure.
This will replace diff-to-changed-lines.awk, and also provide a basis
for working with patches, and doing incremental operations,
and interactive applications.

Parser is capable of consuming/preserving/ignoring metadata lines.

This first cut primarily tests parsing and unparsing (reversibility).

A lot of substructure is designated "internal" for the moment, but
subject to future reconsideration.

PiperOrigin-RevId: 317222088