interactive patching: verible-patch-tool apply-pick PATCHFILE
Apply this hunk? [y,n,q,?]

This modifies files (referenced in the patch file) in-place.
This tool is transformation agnostic, the patch could result from a
formatter tool like verible-verilog-format, or sed/awk.

It is expected that the affected files are present on the filesystem
at the paths indicated in the patch.  Any needed path adjustments will
be in follow-up changes.
This should hold true if the formatter tool and this program are
invoked from the same working directory, which will be done at the next
level of script integration.

PiperOrigin-RevId: 319466202