Script to format workspace incrementally and interactively.
usage: verible-verilog-format-changed-lines-interactive.sh
(run from p4 workspace)

This script is to be installed along side the other binaries and scripts.

verible-transform-interactive.sh: updated to take a --per-file-transform-flags
argument which is where one can inject a command to compute the set of changed
--lines.  This reduces the patch generated by the formatter to only lines that
were touched relative to the diff-base.

Initial revision works with p4.  git support to follow later.

PiperOrigin-RevId: 322270012