Fold token partition filtering into .extract_alignment_groups.
This refactoring simplifies the TabularAlignTokens() interface.
extract_alignment groups is now expected to produce a filtered array of token
partitions, instead of a contiguous range.  This gives the caller more control than before.

For now, we provide an function-composition adapter 'ExtractAlignmentGroupsAdapter' to reuse existing legacy
functions, and punt having to rewrite them until a follow-up change.

PiperOrigin-RevId: 330579173