Pair AlignmentGroupAction with a subtype tag (int)
... to support handling different alignment groups.

It is up to the caller to assign meaning to the enumerated
values, which are returned by GetPartitionAlignmentSubranges'
partition_selector function.
These values come back in the TaggedTokenPartitionGroups,
so the caller can switch on them and assign different column schema scanners
and policies to each type of alignable group.

No functional change.

PiperOrigin-RevId: 332052172