Alignment adapter interface now can select among multiple choices using subtype
This uses the same subtype enumeration that the was used to tag alignment
groups of token partitions.

This makes it easier to dispatch one of multiple alignment handlers.

Scanner functions and policy functions (from style structure) are now organized
into a library of handlers, keyed on syntax subtype enum.
This will enable reuse of handlers that are applicable to syntax constructs
that can appear in multiple places.

PiperOrigin-RevId: 332297037