First step moving development relevant information to own document.
The toplevel README gets a bit big. This is the first step separating
out information about the use of the project and development resources,
starting with source-code indexing of the C++ code-base.
In particular the Kythe part will get confusing when we add information
how to use Verible to Kythe-index Verilog code-bases, so breaking this
out first will help minimize that confusion.

PiperOrigin-RevId: 332559646