Handle alignment of sections inside ifdefs: port declarations.
Rememdy is to flatten the contents of preprocessor-conditional-balanced blocks,
so that the alignment engine sees all conditionally wrapped port declarations
as alignable rows within the same alignment group.

It was easier to suppress creating a level of token partition tree hierarchy
than to reshape/flatten it after the fact.

The same technique can also apply to other preprocessor-balanced alignable
syntax cases.

fixes #510

PiperOrigin-RevId: 334389134