Fix all global static matcher objects, avoid initialization ordering.
The code is the documentation in this case, best to make the examples valid!
The current state reflects a long lineage of copy-paste-modify for style lint
rules.

This also has the benefit of thinning out included headers in header files,
and moving them into the .cc files.
Lint rule classes don't need to publish that they are using matchers, when that
is just an implementation detail.

Reviewed style-linter developer docs, no changes needed.

PiperOrigin-RevId: 339680068