[CST] Tolerate wrong CST node type on invalid syntax, don't crash.
Recall that in a module/generate item context, the current grammar
implementation tolerates "f()" and rejects ";" (delayed rejection),
when the earliest a parser could report a syntax error is at the "(" in "f();".

Unfortunately, the semantic action of the reduction of "f()" happens before
the rejection of ";", so we let "f()" pass-through as a function call,
expecting it will get rejected.

PiperOrigin-RevId: 341617761