MapTree implementation of trie/prefix-tree data structure.
This will be the basis for hierarchical symbol tables (and subcommand maps).

Many method names and interfaces are modeled after VectorTree.
Like VectorTree, each non-root node links to its unique parent node, so it is possible to navigate up and down.

This round implements the interfaces needed for symbol table construction and
querying.

Next:
* wrapper around this for a navigable symbol table interface.
* populate this symbol table in one pass to collect definitions
  * be able to get VNames and signatures from the nodes
* use the populated symbol table in a second pass to bind symbol references
PiperOrigin-RevId: 347093296