[SymbolTable] ResolveLocallyOnly() for references into the same scope of origin
ResolveLocallyOnly() binds references of symbols that are defined
in the exact same scope.  e.g.

  module m;
    wire w;
    foo bar(w);  // 'w' is resolved, but 'foo' is not
  endmodule

Why? This reduces the set of unresolved unqualified symbols to account for in
inter-file dependency analysis.  This can be significant because large modules
can easily have 1000s of local nets and signals.
This leaves unresolved type references (of which most are non-local), and other
global level references for dependency analysis.

issues #574

PiperOrigin-RevId: 350270583