[SymbolTable] out-of-line function declarations
e.g.
  class base;
    extern function int func();
  endclass
  function int base::func();
  endfunction

References to "base" of out-of-line definitions must be resolved at Build()
time, i.e. base will not be created out of nowhere, but "func" will currently
inject a definition (symbol table node) into "base" if it doesn't already exist.
Both prototype and header should point to the same symbol table node.

Port names are part of the public interface, and thus are officially declared
as part of the prototype, and expected to be the same in the out-of-line
definition (but not verified at this time, LRM 8.24).

Includes minor refactoring of symbol lookup/resolution.

PiperOrigin-RevId: 350272110