Dismabiguate 'with' in stream expression in lexer.
The stream expression has an expression followed by
an optional 'with' '[' array_range_expression ']'

There are a few other cases where an expression can
be trailed with 'with',
 * with_constraint_block: there we expect an
   identifier list in parenthesis after 'with'
 * similar with array methods with predicate
   with '(' ...)
(see previous commit that highlighted these cases).

Since the parser with one look-ahead can't see beyond
the 'with', it runs into a shift/reduce conflict
as it does not know if '(' or '[' is to follow.

Disambiguate that in the lexer and make
TK_with_followed_by_bracket a separate token.

Signed-off-by: Henner Zeller <h.zeller@acm.org>