Work in progress: fix #693
The stream expression has an expression followed by
an optional 'with' '[' array_range_expression ']'
(LRM 11.4.14)

There are a few other cases where an expression can
be trailed with 'with',
 * with_constraint_block: there we expect an
   identifier list in parenthesis after 'with'
 * similar with array methods with predicate
   with '(' ...)

Since the parser with one look-ahead can't see beyond
the 'with', it runs into a shift/reduce conflict
as it does not know if '(' or '[' is to follow.

Disambiguate that in the lexer by creating a
separate token TK_with__followed_by_bracket which is
a TK_with, where the lexer already looked ahead and
knows that it will be followed by '['.

After seeing the '[', everything but the "with" is put
back into the stream to be processed separately.

TODO:
  [ ] Extend what can be between 'with' and '[' to
      include arbitrary whitespace (including: comments)
  [ ] Make formatter work; currently
      {<< {unpacked_arr with [i*4 +: 4]}};
      is formatted
      {<<{unpacked_arrwith[i*4+:4]}};

Signed-off-by: Henner Zeller <h.zeller@acm.org>