Provide more detailed choices for linter autofix.
The autofix choices were easy to set in a way that
it would accidentally overwrite the input file (setting
--autofix=yes and forgetting output file). This patch
makes the choice explicit and in particular requires
to choose 'inplace' option to modify output.

  * no
  * patch-interactive
  * patch
  * inplace-interactive
  * inplace

The `yes` option does not exist anymore, as it is replaced
by the more detailed choices. Hopefully that will not break
too many users at this point. The github action though
is probably affected.

While at it, make the default output stdout for the patch
choices (and move error messages to stderr). That way, the
tool is behaving like other command line tools.

Signed-off-by: Henner Zeller <hzeller@google.com>