Add a lint rule that detects truncated number literals.
Detects number literals that have a shorter bit-width than the
digits provided, e.g. 16'h1FFFF.

Decimal numbers are only dealt with correctly up to 128 bits,
beyond that, only some lower-bound estimates are applied so
it might miss complain opportunity there. Given that super-long
constants are very likely not given as decimal, this
tradeoff is accepted.

Be helpful in error messages in cases where we can only assume
a lower bound; not only in the decimal number heuristics, but
as well in 'don't care' situations: 4'hzz uses _at least_ 5 bits,
so distinguish that in the error message from 4'h1F with uses
_exactly_ 5 bits.

Signed-off-by: Henner Zeller <hzeller@google.com>