Have lint rules only return one descriptor struct.
Previously, we had Name() and GetDescription() in
each lint rule, this is replacing this with one
call GetDescriptor() that returns a struct.

This now also allows to have the configuration options
described in that meta-data.

The result is a lot less boilerplate and more readable
descriptions of lint rules.

Did away with the different formatting styles in the
descriptor (markdown vs. text), as markdown is sufficiently
readable text anyway. There are still a few remnants that need
to be cleaned up, but this is a good first start.

Manually verified that the generated markdown description
is essentially the same.

Signed-off-by: Henner Zeller <hzeller@google.com>