Provide an additional alternative for autofixing zero literals.
Autofixes for undersized zero literals, e.g. 32'h0 provide
as one alternative to fix it to a simple '0. Thus the alternatives
for a zero constant `32'h0` are now

  1. `'0`
  2. `32'h00000000`
  3. `32'd0`

Also: move the decimal alternative to always be after the zero
expansion, to avoid changing the type.

That also means we don't need this anymore as a configuration choice,
making things simpler.

Signed-off-by: Henner Zeller <hzeller@google.com>