Provide UI choice if there are multiple autofix alternatives.
Most lint rules with autofix have exactly one choice of fixes,
but there are some that have 2 or even three alternatives.

Provide a way to choose if there are multiple options.
The short choices are then including the digits
'1', '2', etc. The default 'y' is equivalent to choosing
choice 1.

If there are multiple options, print all of these and
mark as "1. alternative", "2. alternative", etc.

Also while at it: improve interactive sessions by always
printing the available patch first (or patches, if there are
alternatives), as the user typically will not
choose blindly, so make the initial 'print' automatic.

Signed-off-by: Henner Zeller <hzeller@google.com>