Provide way to process all changed buffers since last time we looked at them.
For running more expensive operations such as linting, it would be
nice to just look at buffers that have changed since last time we looked
at them.

Introduce the concept of a 'global version' which is counted for
the whole buffer collection and increments whenever some change happens
on any of the managed buffers, or a new buffer has been added.

Provide a MapBuffersChangedSince() method that calls a function for
only the buffers that have changed since a particular version number.

Thus a task that regularly looks at buffers to work on can do this
efficiently by only remembering a single number.

Signed-off-by: Henner Zeller <h.zeller@acm.org>