Add a simple file event dispatcher.
A wrapper around select() to be able to wait on a file descriptor
and execute something or on timeouts.

API allows to register callback functions for both cases.

We needs this in the LSP implementation to listen on stdin,
from which we get commands as well have to run regular tasks such
as linting when there is a gap in the chatter coming from the
client.

This uses standard select() underneath which is the lowest
common denomiator on any unix platform but it might need a different
implementation on other platforms. For now, #ifdef-ing things
out that won't compile to be implemented later.

Signed-off-by: Henner Zeller <h.zeller@acm.org>