Allow for lenient header parsing with simple newline.
The header requires to have CRLF line endings, so for
end of headers that would be two of these: \r\n\r\n.

For 'manual' speaking of the protocol on the shell it would
be desirable to have a 'lenient' mode that accepts
simple newlines. Add this as a possible mode with
a new default parameter to the constructor
"strict_crlf_header_separation" (default true).

Signed-off-by: Henner Zeller <h.zeller@acm.org>