Provide function to extract error information from a rejected token.
This functionality is used in two places: to build a string to
print errors and in the extraction of the json output for
verible-verilog-syntax.
In the future, there will be another use when generating the
information for the language server.

This function makes it simpler to do custom formats and limits
the leaky abstractions accessed by the user, such as
TextStructure line column map; for the user they are
interested in getting ready resolved line/column-Range, so this
is readily provided.

To convey the full information, introduce a LineColumnRange
to represent the start and end of the region to complain about.
Add printing functionality to print it in a way compatible with
common editors <line-start>:<coli-start>:<line-end>:<col-end>:

The printing is _not_ yet used to first make sure other tools
relying on the format don't break; the finaly switch is then
one simple change (TODO left in code).

Use the new function in

  * printing an syntax error message for the user
  * generate the JSON diagnostic print.

Signed-off-by: Henner Zeller <h.zeller@acm.org>