Refine the smoke test: test on multiple projects.
Run the smoke test (all verible binaries on all files)
on a few projects that can be readily checked out from
github, this will help uncover regressions quickly as
it is run on 'real code'.

There are a few projects that are _not_ included
currently as they crash the formatter. These have
all been recorded in
issues #1004 #1005 #1006 #1007 #1008 #1010 #1011 #1012

The verible-verilog-project tool is currently not run
because of issues #917 #1002 #1003

Signed-off-by: Henner Zeller <h.zeller@acm.org>