Language server: Publish Diagnostics.
Diagnostics found in the parsed file (syntax
or linting) are published as a Notification back
to the client.

Introduce a verible-lsp-adapter set of utility
functions (right now: exactly one), that simply converts
the internal verible state of syntax error and lint status
in Language Server compatible objects - converting
one struct into another.
Since these are the code-generated structs, these then
can be automatically serialized to json.

The verible-verilog-ls_test.sh walks through all the
functionality and looking at the returned diagnostics.
In particular we test what happens if the document is
edited to fix an issue - we get a diagnostic back that
is empty. This test walks all code-paths so gives us
100% coverage.

Signed-off-by: Henner Zeller <h.zeller@acm.org>