Language Server: Add autofix code action.
This reacts to the textDocument/codeAction JSON RPC
request and responds with possible fixes in the requested
region. Adds one more translation function in
verible-lsp-adapter that converts verible internal
structs to LSP external structs.
Tested in verible-verilog-ls_test by requesting and
edit for the (autofix available) no-newline-at-EOF lint rule.

(also tested with a bunch of editors (emacs, sublime, kate)
that it behaves as expected)

Signed-off-by: Henner Zeller <h.zeller@acm.org>