Use bazel cache again - the ccache sees too many non-cacheable commandline changes.
After experimenting #1033 turns out that using ccache would require more work
in the CI (it works locally, but I suspect that re-creating the build environment
in each pull request creates situations that things can not be cached).

This undoes #1032 but slightly renames the cache to get a fresh clean
slate to get the cache unstuck.

(bazel action is still deprecated and they suggest basilisk for
possibly things that go beyond caching. So we still should probably
implement bazel caching but in a way that is similar to how the
ccache action is implemented with timestamped versions).

Signed-off-by: Henner Zeller <h.zeller@acm.org>