Output token columns to be _character_ not _byte_-offset in line.
Take multi-byte characters into account.

Remove the non-descript operator() in LineColumnMap and replace
with GetLineColAtOffset(). Also add such a convenience method
in text view to reduce abstraction leakage. Ideally, we would
actually never have to deal with 'raw' LineColumnMaps inside
code, only with attached text view.

The GetLineColAtOffset() needs the original text to correctly
count UTF8-characters in the line in question; this should
probably be replaced with either
  * remembering the original string-view from constructing
    the LineColumnMap (but CAVE: lifetime issues)
  * or maybe a sub-structure that remembers multi-byte
    offset within a line.
For now, this is not a big deal as the original string-view
always is available.

Fixes #1047 #1059

Signed-off-by: Henner Zeller <hzeller@google.com>