Cache: improve Bazel caching and also use for for macos.
Always store caches with the latest timestamp and have them find
in reverse order from the internal cache store.
This works around the one-time cache writing of the actions/cache
action and provides the feature we want: use the latest known cache
as a starting point for a new build.

Caching for windows is added in the CI, but there seems to be a
permission problem preventing actually storing it. To be fixed
later.

To reduce cache sizes in attempt to stay below the available 5GiB
space, make the compile actions just compile the :install-binaries

Signed-off-by: Henner Zeller <h.zeller@acm.org>