Disable use of cache in Windows for now
The Windows cache is using the most of all build caches
(over 3GiB), but it is also leas effective as the packaging of
the cache takes about 45 min vs. a 55min fresh build.

These 3GiB will then evict other caches so that
we have to re-build a lot of other parts of the CI: the
total cache size allowance is currently 5GiB (see [cache limits]).

So for now, let's disable it, but keep it prepared for a future
in which there is more space available and possibly the more
efficient zstd compression is used also on Windows.

[cache limits]: https://docs.github.com/en/actions/advanced-guides/caching-dependencies-to-speed-up-workflows#usage-limits-and-eviction-policy

Related: https://github.com/actions/cache/issues/442
         https://github.com/actions/cache/issues/301

Signed-off-by: Henner Zeller <h.zeller@acm.org>