Move FormatStyle flag initialization into separate library.
Having it a separate library allows to use it also in other
binaries (such as the language server). Also this provides
a context to implement other initialization from a configuration
file.

This does _not_ change or add any flags, but this is in
preparation to provide additional flags for previously
not flag-configured fields.

Signed-off-by: Henner Zeller <h.zeller@acm.org>