Add FormatVerilogRange() formatting functions that only formats and emits a range of lines.
This is needed for use in the language server that generally only
requests a user-requested sub-set of lines to be formatted.

We can't just use the regular FormatVerilog() with a sub-range,
as it is not possible to determine which of the formatted output
lines correspond to the original region, hence we need a function
that _only_ outputs the code from the original region.

Signed-off-by: Henner Zeller <h.zeller@acm.org>