Add full-file formatting in the language server.
For this formatting option, the regular FormatVerilog() is
used that does a validation step over the whole file to make
sure the file is not messed-up accidentally.

For that, factor out the internal part of FormatVerilog() that
takes a text structure (which is what we have available in the
language server) and writes to a string; this is then used in
the original FormatVerilog() that does additional checks.

Signed-off-by: Henner Zeller <h.zeller@acm.org>