Fix some clang-tidy warnings found by readability-qualified-auto.
Also organized the .clang-tidy rules to sort disabled rules.

Unfortunately, we can't have readability-qualified-auto on by
default, as it also suggests to qualify the return value of
absl::string_view::begin() with const auto* - which can be wrong
as absl::string_view might be typedefed to std::string_view whose
iterator type is implementation defined. So this change does
not touch any suggestions that involves iterators.

Signed-off-by: Henner Zeller <h.zeller@acm.org>