For the Kythe result, always stream the result.
Instead of collecting the final result in an ever-growing set
before printing, write them out as they are received to
minimize memory use.

This _could_ result in the same facts be emitted multiple times
if there are duplicates (can there ?), but this is benign.

Make StreamKytheFactsEntries() the main toplevel function.
Remove the ExtractKytheFacts() that accumulates everything in
sets first.

Signed-off-by: Henner Zeller <hzeller@google.com>