Prepare users of VectorTree to be more agnostic about the internal container.
This will make it simpler to change the container later. In particular,
not all containers might have reserve(), so pull that operation into a
method that can be implemented according to the containers' needs.

Also, make the `subnodes_type` typedef in VectorTree public for users
to properly define their iterator types.

This is a preparation for other changes to be simpler, so
no functional or container type change in this PR.

Signed-off-by: Henner Zeller <h.zeller@acm.org>