Improve KeywordIdentifier detection in symbol table using IsIdentifierLike()
The overlapping keywords and identifiers are the result of
mis-classification in the SymbolTable by simply looking
at SymbolIdentifier; instead, use the IsIdentifierLike()
predicate function.

Also, while at it, reduce the number of special cases by
only keep the KeywordIdentifier tokens that are also used
in other places of the grammar.

Long term, we need to make sure to properly handle
KeywordIdentifiers in the whole stack; it seems like there
are other subtle bugs lurking.

One approach would be to change the token tag to
SymbolIdentifier if we happen to encounter it in
KeywordIdentifiers, but currently there are hidden assumptions
that this is not happening.

Fixes #917

Signed-off-by: Henner Zeller <hzeller@google.com>