Add `ifdef/`elsif/`else preprocessing branch filtering.
This is a configuration option for the preprocessor, and currently
is not enabled by default.
We do not necessarily want it to be enabled for the formatter in any
case, but this might be useful for the syntax checker or linter (not
enabled in this change).

Implementation: keep a stack of `BranchBlock`s that represent nested
if conditions. Evaluate existence of macro-names used to choose the
branch from the internal macro definitions.

TODO:
  - provide a way to inject macro definitions from the
    command line or project-filelist (+define+)
  - Possibly leaning towards always  enabling for kythe and project tool.
  - The following tools might benefit from enabled branches, but also
    from testing permutations:
  - add branch processing to the syntax checker and linter; possibly
    off by default and then auto-fallback if the first parse couldn't
    complete unless otherwise chosen.
  - Language server: always enable.

Precondition for issues #1118 #377 #241 #228 #183
Signed-off-by: Henner Zeller <h.zeller@acm.org>