Surface the verilog preprocessor config up one level.
Add VerilogPreprocess::Config to all analyzer convenience
methods, such as AnalyzeAutomaticMode() but as well as all
the excerpt parse methods, such as AnalyzeVerilogExpression(...).

Wire this up all the way to the tools where they are used (and tests),
but choose the safe preprocessor choice defaults for now, to be
changed later.

So no functional change right now, just preparation to make it
possible to choose configuration.

Signed-off-by: Henner Zeller <h.zeller@acm.org>