Document for each project the expected number of tool failures.
Tools fail on some of the files. So far, we only were concerned about
crashes in tool in the smoke tests, but recently improvements have
been sen in also reducing the number of files we can't process.

This change documents how many of these failures are to be expected
to make sure we don't add regressions (if failures increase, the
smoke test will fail), but also detect when we see less than expected
failures.

Signed-off-by: Henner Zeller <h.zeller@acm.org>