Update googletest to latest.
This unfortunately also means, that it pulls in a dependency to RE2.

And this RE2 does not work with absl::string_view (unless
absl::string_view is a typedef for std::string_view). So
had to patch RE2 to accept that string-view (will attempt to
upstream that to RE2; first attempt
https://github.com/google/re2/pull/385 but looks like I need to
file it via gerrit...)

Also update absl, as googletest does not work with the older one.
... which in turn requires another "platforms" dependency if compiled with
older thatn bazel 5. :/

Signed-off-by: Henner Zeller <h.zeller@acm.org>