Unbutu releases: also link libstc++ statically for newer distributions.
Similar to the xenial and bionic distributions (that require a different
g++), also use the static libstdc++ linking for releases of the newer
ubuntus. That way we are slighly more resilient to library version
mismatch in case the download is attempted to run in
Debian distributions/releases.

Downside: larger release tar to download.

Signed-off-by: Henner Zeller <h.zeller@acm.org>