Make InitCommandLine() return vector of string_view instead of char*
For all intendt and purposes, we're only interested in string_view
for the positional parameters and in particular do not want to encourage
modifying them (in particular since all kind of split operations can
cheaply and easily be done in the string_view domain).

The `char*` was just a leaked implementation detail of the
underlying main() entry point which happens to return a writable
buffer.

Signed-off-by: Henner Zeller <hzeller@google.com>