We only need filelist as project dependency in filelist writer.
That reduces the project-specific dependencies significantly:

$ bazel query 'deps(//verilog/tools/kythe:verilog-kythe-kzip-writer)' | grep -v '^@' | wc -l
    246
$ bazel query 'deps(//verilog/tools/kythe:verilog-kythe-kzip-writer)' | grep -v '^@' | wc -l
     37

Unfortunately, this does not result in a binary size reduction for the
stripped binary. In both cases the binary is about the same size (and
one of the larger binaries in verible), which is probably because the
linker works well in figuring out cruft.
Maybe the size comes from the additional libraries, such as proto buffer
and crypto that it is so bloated ?

Signed-off-by: Henner Zeller <h.zeller@acm.org>