Change functions parsing file list from return to append.
We want to merge parameters from the command line and files
in file-lists and accumulate them into an effective filelist
(possibly with multiple `-f` parameters, interspersed with
commandline parameters ?).

Change the parse functions to append to an existing file list,
to make it easy programmatically to do the accumulation.
Remove the FileList::Append() method.

Signed-off-by: Henner Zeller <hzeller@google.com>